requirements.txt
schedule~=1.1.0
influxdb~=5.3.1
requests~=2.28.1
python-dateutil~=2.8.2